example/requirements.txt
fastapi
pydantic
uvicorn
Pillow
